Impl 3rd kaleido tut; use variants, impl Visitor, Driver
Impl the 3rd part of the kaleidoscope tut by adding the IR
creation from the AST.

Impl `Visitor` class which is used in conjunction with `std::variant`s
to navigate the AST and produce IR.

Impl output the result of the IR generation if the debug mode is high
enough.

Apply other general fixes and adjustements.